Post-review fixes
