+++ js/utils/addBackButtonAction.js
@@ -1,0 +1,6 @@
+import {renderScreen as renderGreetingScreen} from '../screens/greeting';
+
+export default (renderedScreen) => {
+  const backButton = renderedScreen.querySelector(`.back`);
+  backButton.addEventListener(`click`, () => renderGreetingScreen());
+};
